+++ shims/spark310/src/main/scala/com/nvidia/spark/rapids/shims/spark310/ParquetCachedBatchSerializer.scala
@@ -16,9 +16,8 @@
 
 package com.nvidia.spark.rapids.shims.spark310
 
-import java.io.{ByteArrayOutputStream, File, InputStream}
+import java.io.InputStream
 import java.nio.ByteBuffer
-import java.nio.file.Files
 
 import scala.collection.JavaConverters._
 import scala.collection.mutable
@@ -28,9 +27,15 @@ import ai.rapids.cudf._
 import com.nvidia.spark.rapids._
 import com.nvidia.spark.rapids.GpuColumnVector.GpuColumnarBatchBuilder
 import com.nvidia.spark.rapids.RapidsPluginImplicits._
+import org.apache.commons.io.output.ByteArrayOutputStream
 import org.apache.hadoop.conf.Configuration
+import org.apache.hadoop.mapreduce.RecordWriter
 import org.apache.parquet.HadoopReadOptions
-import org.apache.parquet.hadoop.{ParquetFileReader, ParquetInputFormat, ParquetOutputFileFormat}
+import org.apache.parquet.column.ParquetProperties
+import org.apache.parquet.hadoop.{CodecFactory, MemoryManager, ParquetFileReader, ParquetFileWriter, ParquetInputFormat, ParquetOutputFormat, ParquetRecordWriter, ParquetWriter}
+import org.apache.parquet.hadoop.ParquetFileWriter.Mode
+import org.apache.parquet.hadoop.api.WriteSupport
+import org.apache.parquet.hadoop.metadata.CompressionCodecName
 import org.apache.parquet.io.{DelegatingPositionOutputStream, DelegatingSeekableInputStream, InputFile, OutputFile, PositionOutputStream, SeekableInputStream}
 
 import org.apache.spark.TaskContext
@@ -121,12 +126,12 @@ class ByteArrayInputFile(buff: Array[Byte]) extends InputFile {
   }
 }
 
-class ByteArrayOutputFile(stream: ByteArrayOutputStream) extends OutputFile {
-  // default to 32 MB: large enough to minimize the impact of seeks
-  var blockSize = 32L * 1024 * 1024
+object ByteArrayOutputFile {
+  val BLOCK_SIZE = 32 * 1024 * 1024 // 32M
+}
 
+class ByteArrayOutputFile(stream: ByteArrayOutputStream) extends OutputFile {
   override def create(blockSizeHint: Long): PositionOutputStream = {
-    blockSize = blockSizeHint
     new DelegatingPositionOutputStream(stream) {
       var pos = 0
       override def getPos: Long = pos
@@ -153,7 +158,7 @@ class ByteArrayOutputFile(stream: ByteArrayOutputStream) extends OutputFile {
 
   override def supportsBlockSize(): Boolean = true
 
-  override def defaultBlockSize(): Long = blockSize
+  override def defaultBlockSize(): Long = ByteArrayOutputFile.BLOCK_SIZE
 }
 
 class ParquetBufferConsumer(val numRows: Int) extends HostBufferConsumer with AutoCloseable {
@@ -557,27 +562,32 @@ class ParquetCachedBatchSerializer extends CachedBatchSerializer with Arm {
     } else {
       // fallback to the CPU
       input.mapPartitions {
-        cbIter =>
+        internalRowIter =>
           new Iterator[CachedBatch]() {
             private val parquetOutputFormat =
               SQLConf.withExistingConf(broadcastedConf.value) {
                 new ParquetOutputFileFormat()
               }
 
-            override def hasNext: Boolean = cbIter.hasNext
+            override def hasNext: Boolean = internalRowIter.hasNext
 
             override def next(): CachedBatch = {
               // Each partition will be a single parquet file
               var rows = 0
-              val stream = new ByteArrayOutputStream()
+              // at least a single block
+              val stream = new ByteArrayOutputStream(ByteArrayOutputFile.BLOCK_SIZE)
               val outputFile: OutputFile = new ByteArrayOutputFile(stream)
 
               val recordWriter = parquetOutputFormat.getRecordWriter(outputFile,
                 broadcastedHadoopConf.value.value)
 
-              while (cbIter.hasNext) {
+              while (internalRowIter.hasNext) {
                 rows += 1
-                val row = cbIter.next()
+                if (rows < 0) {
+                  throw new IllegalStateException("CachedBatch doesn't support rows larger " +
+                    "than Int.MaxValue")
+                }
+                val row = internalRowIter.next()
                 recordWriter.write(null, row)
               }
               // passing null as context isn't used in this method
@@ -593,5 +603,45 @@ class ParquetCachedBatchSerializer extends CachedBatchSerializer with Arm {
      cachedAttributes: Seq[Attribute]): (Int, Iterator[CachedBatch]) => Iterator[CachedBatch] = {
     //essentially a noop
     (partId: Int, b: Iterator[CachedBatch]) => b
+  }
+}
+
+/**
+ * Similar to ParquetFileFormat
+ */
+class ParquetOutputFileFormat {
+
+  val writeSupport = new ParquetWriteSupport().asInstanceOf[WriteSupport[InternalRow]]
+  val DEFAULT_MEMORY_POOL_RATIO: Float = 0.95f
+  val DEFAULT_MIN_MEMORY_ALLOCATION: Long = 1 * 1024 * 1024 // 1MB
+
+  def getRecordWriter(output: OutputFile, conf: Configuration): RecordWriter[Void, InternalRow] = {
+    import ParquetOutputFormat._
+
+    val blockSize = getLongBlockSize(conf)
+    val maxPaddingSize =
+      conf.getInt(MAX_PADDING_BYTES, ParquetWriter.MAX_PADDING_SIZE_DEFAULT)
+    val validating = getValidation(conf)
+
+    val init = writeSupport.init(conf)
+    val w = new ParquetFileWriter(output, init.getSchema,
+      Mode.CREATE, blockSize, maxPaddingSize)
+    w.start()
+
+    val maxLoad = conf.getFloat(MEMORY_POOL_RATIO, DEFAULT_MEMORY_POOL_RATIO)
+    val minAllocation = conf.getLong(MIN_MEMORY_ALLOCATION, DEFAULT_MIN_MEMORY_ALLOCATION)
+
+    lazy val memoryManager = new MemoryManager(maxLoad, minAllocation)
+
+    val writerVersion =
+      ParquetProperties.WriterVersion.fromString(conf.get(ParquetOutputFormat.WRITER_VERSION,
+        ParquetProperties.WriterVersion.PARQUET_1_0.toString))
+
+    val codecFactory = new CodecFactory(conf, getPageSize(conf))
+
+    new ParquetRecordWriter[InternalRow](w, writeSupport, init.getSchema, init.getExtraMetaData,
+      blockSize, getPageSize(conf), codecFactory.getCompressor(CompressionCodecName.UNCOMPRESSED),
+      getDictionaryPageSize(conf), getEnableDictionary(conf), validating, writerVersion,
+      memoryManager)
   }
 }
